Move files for cr
