+++ apps/spruce/src/components/Settings/EventLog/EventDiffTable/utils/keyRenderer/KeyRenderer.test.tsx
