+++ apps/spruce/src/components/Settings/EventLog/EventDiffTable/utils/keyRenderer/index.ts
